Latest changes

+++ alembic/versions/005.py
@@ -11,8 +11,8 @@ import sqlalchemy.types
 
 def upgrade():
     op.create_table('setting', 
-        sa.Column('id',         sa.Integer(11), nullable=False),
-        sa.Column('user_id',    sa.Integer(11), nullable=False),
+        sa.Column('id',         sa.Integer, nullable=False),
+        sa.Column('user_id',    sa.Integer, nullable=False),
         sa.Column('key',        sa.Unicode(100), nullable=False),
         sa.Column('value',      sqlalchemy.types.BLOB, nullable=False),
         sa.Column('created_at', sa.DateTime, nullable=False),
